Refactored event emitter common code into utility
