+++ packages/api-utility/index.ts
@@ -1,5 +1,7 @@
 import Uri from './src/uri';
+import Emitter from './src/emitter';
 
 export {
-  Uri
+  Uri,
+  Emitter
 };
